pyproject.toml
[tool.poetry]
name = "space-station"
version = "0.1.0"
description = ""
authors = ["Mikita Yeudakimau <nikevdokimovv@gmail.com>"]
readme = "README.md"
#packages = [{include = "space_station"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
drf-spectacular = "^0.24.2"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
